Add a section on capability control registers.
This idea was somewhat defined previously for KCC and KSC and then
reused for CSTAR and KGS, but this arranagement was a bit awkward,
especially once additional references were added for new CHERI
instructions and ASR.  Instead, define the idea of a capability
control register, its restrictions, and how it is accessed in a new
section.  KCC, KSC, CSTAR, and KGS are then defined as capability
control registers which simplifies the containing sections.

While here, add some additional cross references and chop some
excessive words.