chap-architecture: Fix alignment requirements specification
There is an edge case (that has already been discussed in the algorithm
above) when we must add 1 to E due to the rounded-up length overflowing
and causing the top two bits of T to no longer be 01. A CHERI-128
implementation of this property SMT checks against the full Sail
implementation.