cheri-x86-64: Change the opcode prefixes for capability mode.
Specifically, to more closely match the 0x66/0x67 prefixes used for 16
vs 32 bit mode in 32-bit x86, use 0x06 to toggle the operand size
(similar to 0x66), and 0x07 to toggle the addressing mode (similar to
0x67).  The previous prefixes were 0x62 for toggling the addressing
mode, and 0x07 for toggling the operand size.  0x16/0x17 is another
pair of invalid opcodes in 64-bit mode that could also be used to
provide similar symmetry relative to 0x66/0x67.