Rework table macros to make a bit more sense
Passing in the class only to feed it back to func2str as an additional
argument doesn't really make any sense, is a bit hard to follow and
bakes in an assumption about how encodings are used (that there is a
flat namespace; some uses might not even want a class if they just have
a single encoding table). Instead add RISC-V-specific wrapper macros
for each class.

Whilst here reduce some code duplication.