RISC-V chapter: Remove references to CHERI-MIPS and make standalone.
I have not removed all mention of split register files as done in
other chapters.  It might be worth revisiting some of the parameters
such as register files and opcode encoding choices and deciding if
some of the alternates should be removed (or perhaps relegated to
items in the rationale chapter) to streamline RISC-V spec.