x86: Further simplify additional capability registers.
With the previous change to reuse FS: and GS: segment prefixes to
select CFS and CGS as the base capability for memory operands, there
is no longer a hard requirement to permit access to additional
capability registers as direct operands in most instructions.  The
typical use cases would be to only move values into and out of
additional capability registers similar to SCRs on RISC-V.

Based on that, relegate additional capablity registers to an isolated
bank similar to control and debug registers only accessible by moving
to/from a capability GPR.

There is an extra wrinkle that some of the additional capability
registers should be accessible by user mode whereas others should be
restricted to kernel mode only.  For now the sketch suggests that the
relevant MOV opcodes will determine the required privilege level based
on the register index.  Other possibilities are enumerated in the
rationale section.