x86: Add instruction descriptions for PUSH and POP.
These instructions default to capability operands in capability mode
and some operand sizes available in 64-bit mode are not available in
capability mode.  To try to clarify this, these tables include entries
for instructions available in other modes but not capability mode.
This even includes a few variants available in neither 64-bit nor
capability mode.