x86: Suggest additional encodings for arithmetic instructions.
As a two-operand ISA, arithmetic instructions necessarily overwrite
one of the input operands with the arithmetic result.  In some cases,
e.g. when adding various offsets to a single base pointer in a
sequence, it may be convenient to overwrite the offset operand with a
capability result rather than the input capability.

Propose a scheme where the existing 8-bit arithemtic operations would
be extended via the capability prefix to support a pseudo-three-way
operand mode where a single register or memory location is used as
both an integer source operand and capability destination operand.