RISC-V: Remove DDC/PCC offsetting entirely for now.
In the interest of having something concrete to ship in Version 9,
remove DDC/PCC offsetting entirely.  This includes removal of the
unaligned PCC base exception and changing the description of SCRs that
extend CSRs such that the CSR now reads/writes the address of the SCR
rather than the offset.

A new rationale subsection does leave room for adding back a form of
relocation for user mode in the future.