RISC-V: Trim language around the CHERI ISA being a design space.
CHERI-RISC-V's definition is more concrete with fewer alternatives
than when the specification was first written.  The only remaining
design space choice is 32-bit vs 64-bit which is probably not enough
to justify the remaining language.  This does still keep the idea of a
design space in the list of goals motivating the initial desire to
apply CHERI to RISC-V as that list is largely historical.